requirements.txt
psycopg2==2.9.5
pytz==2022.6
python-dotenv==0.21.0
APScheduler==3.9.1
fastapi==0.85.2